Rename `action` variable in /user/login.html
For some reason, setting a variable named `action` in this location causes the variable to be set to `""`.
See: https://github.com/NaturalHistoryMuseum/ckanext-ldap/issues/36
I think it is a scoping thing. Perhaps there is another variable named "action" somewhere in the template hierarchy that is set to `""` and overrides this one.

Renaming the variable fixes the problem.

I don't believe this variable is used anywhere else, so renaming it here shouldn't cause any other problems to crop up.