Add the parameter ckanet.ldap.base_dn_alt.
Adds to the description the alternative base Distinguished Name (DN) in case the forest has two Organizational Units (OU) to look for. The way the python-ldap package is built, it needs the DN to have an OU. The original ldap library in C does not need it. Although, as a corner case solution, a new parameter of the alternative DN was created.