Include the parameter ckanext.ldap.base_dn_alt.
Include the alternative Distinguished Name (dn) parameter. It seems (needs confirmation) like the way the python-ldap package was built, it does not accept a DN wit only the Domain Component (DC), it must have a OU. So the parameter ckanext.ldap.base_dn_alt will be useful in case the Active Directory (AD) forest has two Organizational Units. Although, the best solution would be changing the python-ldap package so that the search functions worked with a DN without an OU.