Added new parameter ckanext.ldap.base_dn_alt
This modification permits the use of two base Distinguished Names that will be searched for. It works along with the ckanext.ldap.search_alt, using the same logic. It there was already a way to use two Organizational Units in the cnakext.ldap.base_dn this change would be dull.