Prevent an exception being thrown due to context reuse
If you reuse a context dict after it is used on a failed call you may
get an exception like this: "Action function <action> did not call its
auth function". This is what was happening here where we used the action
"organization_show" to check if the organisation already existed or not
and then if it didn't, create it. The context is reused so the auth
check exception was thrown. Simply removing the auth audit solves the
problem. This PR on ckanext-harvest exhibits the same problem:
https://github.com/ckan/ckanext-harvest/issues/84 and it is fixed in the
same way as we do in this commit, in this commit on ckanext-harvest:
https://github.com/ckan/ckanext-harvest/commit/f315f41c86cbde4a49ef869b6993598f8cb11e2d.